Build native arm64, and fix the crashes that hid behind Rosetta
Installing arm64 SDL2 made lldb and hardware watchpoints usable, which found
three separate memory bugs in a couple of hours.  Every scene that was
crashing now survives 8 runs out of 8.

── Build ──
Makefile.pc took whichever sdl2-config came first in PATH.  On a Mac with
both Homebrews that was the Intel one, so an M1 built x86_64 and ran the
whole game under Rosetta.  It now picks the sdl2-config whose libSDL2.dylib
matches `uname -m`; SDL2_CONFIG= overrides it.  On Linux, where lipo does not
exist, the probe finds nothing and falls back to PATH exactly as before.

── globalVariable was u32 ──
Beatscript opcode 0xB0 (call_result) jumps to whatever the last script
function returned, so D_030053c0.globalVariable carries script addresses —
but it, localVariables[] and the BsFunc typedefs' return type were all 32-bit.
drum_studio_drumming jumped to 0x008d41e0 instead of 0x1008d41e0 on its first
call_result, every single run.  Found with a watchpoint on currentCmd.

── lineStrings was allocated at half size ──
    textPrinter->lineStrings = mem_heap_alloc_id(memID, totalLines * sizeof(u32));
lineStrings is char **.  The loop right below it writes totalLines pointers,
so every text printer overflowed its block by totalLines * 4 bytes.  In the
text scenes that landed on a scheduled-function task, whose function pointer
was then called: a jump to 0x21a0002, reproducible under lldb.

── Scene/engine/cue sizes are GBA sizes ──
Scene.requiredMemory, GameEngine.gameDataSize and CueDefinition.cueInfoSize
are struct sizes recorded when a pointer was four bytes.  struct
DataRoomSceneData is 0x20 there and 0x28 here, so the scene overran its own
allocation by eight bytes the moment it stored its listbox pointer.  The new
GBA_STRUCT_BYTES() in global.h doubles them, which is an upper bound: each
4-byte slot can at most become 8.  Regenerating the tables from real sizes
would be tidier, but they are decompiled data and this is the layer that
knows about the host.

── Results ──
Eight runs each, before this commit -> after:

  drum_studio_drumming  8/8 crash -> 0/8
  data_room             6/8       -> 0/8
  reading               4/8       -> 0/8
  karate_man            0/8       -> 0/8

A 24-scene colour sweep is unchanged except that the two scenes which used to
die during boot now render: data_room 19 colours, drum_studio_drumming 18.

OPT now defaults to -O2.  It was held at -O0 in the previous commit because
the corruption above was outstanding; -O2 did not cause it but did change the
heap layout enough to turn an intermittent crash into a frequent one, which
is how it was found.  -O2 is now 0 crashes in 48 runs with an identical
colour sweep.

Per-frame work over 1240 frames of karate_man, across the whole session:
  x86_64 / Rosetta / -O0   work 7.17 ms  render 6.25  audio 0.80
  arm64  / native  / -O0   work 4.97 ms  render 4.27  audio 0.62
  arm64  / native  / -O2   work 3.16 ms  render 2.91  audio 0.19
against a 16.74 ms budget.

Not fixed, and worth recording: AddressSanitizer is unusable on this host.
Under Rosetta it died with SIGILL; native arm64 it spins forever inside
__asan::InitializeShadowMemory -> get_dyld_hdr, which is Apple clang 17's
bundled runtime against the macOS 26 dyld shared cache.  All three bugs above
were found with lldb watchpoints instead.

+++ include/types.h
@@ -106,9 +106,11 @@ extern struct BeatscriptScene {
     s8_8 musicPitchSrc1;
     s8_8 musicPitchSrc2;
     s8_8 musicPitch;
-    u32 globalVariable; // [D_030053c0 + 0x24] Global (Main Scene) Variable
+    // Pointer-sized, not u32: opcode 0xB0 (call_result) jumps to whatever the
+    // last script function returned, so this slot carries script addresses.
+    uintptr_t globalVariable; // [D_030053c0 + 0x24] Global (Main Scene) Variable
     struct BeatscriptThread threads[2];
-    u32 localVariables[2]; // [D_030053c0 + 0x160] Local (Sub-Scene) Variables
+    uintptr_t localVariables[2]; // [D_030053c0 + 0x160] Local (Sub-Scene) Variables
     s16 unk168;
     u16 unk16A;
     u16 unk16C;
@@ -132,5 +134,5 @@ extern struct BeatscriptScene {
     intptr_t callbackArgument;
 } D_030053c0;
 
-extern u32 *D_03005588; // Current Scene Variable
+extern uintptr_t *D_03005588; // Current Scene Variable
 extern s16 *D_0300558c; // Current Scene Sprite Pool
